any test related to a benchmark should be in the docs
